+++ app/src/main/java/joshuatee/wx/radarcolorpalettes/UtilityColorPalette165.kt
@@ -53,7 +53,7 @@ internal object UtilityColorPalette165 {
         val diff = 10
         dbzList.indices.forEach {
             val lowColor = Color.rgb(redList[it], greenList[it], blueList[it])
-            objectColorPalette.putBytes(redList[it].toByte(), greenList[it].toByte(), blueList[it].toByte())
+            objectColorPalette.putBytes(redList[it], greenList[it], blueList[it])
             (1 until diff).forEach { _ -> objectColorPalette.putInt(lowColor) }
         }
     }
